Game Completed

+++ Assets/Scripts/Enemy Scripts/EnemyHealth.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 
 public class EnemyHealth : MonoBehaviour
 {
-    public float health = 100f;
+    public int health = 100;
     public GameObject deadFX;
 
-    public void TakeDamage(float damageAmount)
+    public void TakeDamage(int damageAmount)
     {
         health -= damageAmount;
         print("Damage");
